Update April 4,2016

+++ table.h
@@ -20,7 +20,7 @@ struct table2D {
   int *axisX16;
   
   //Store the last X and Y coordinates in the table. This is used to make the next check faster
-  byte lastXMax, lastXMin;
+  int lastXMax, lastXMin;
 };
 
 void table2D_setSize(struct table2D targetTable, byte newSize);
